+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ConfigureScrumPokerAction.java
@@ -16,6 +16,7 @@ public class ConfigureScrumPokerAction extends JiraWebActionSupport {
     private static final long serialVersionUID = 1L;
     private static final String PARAM_STORY_POINT_FIELD = "storyPointField";
     private static final String PARAM_SESSION_TIMEOUT = "sessionTimeout";
+    private static final String PARAM_DEFAULT_PROJECT_ACTIVATION = "defaultProjectActivation";
     private static final String PARAM_ACTION = "action";
 
     private final CustomFieldManager customFieldManager;
@@ -42,8 +43,18 @@ public class ConfigureScrumPokerAction extends JiraWebActionSupport {
         return scrumPokerSettingService.loadStoryPointField();
     }
 
+    /**
+     * Current configured session timeout in this Jira instance.
+     */
     public String getSessionTimeout() {
         return scrumPokerSettingService.loadSessionTimeout().toString();
+    }
+
+    /**
+     * Current configured default project activation in this Jira instance.
+     */
+    public String getDefaultProjectActivation() {
+        return String.valueOf(scrumPokerSettingService.loadDefaultProjectActivation());
     }
 
     @Override
@@ -51,8 +62,13 @@ public class ConfigureScrumPokerAction extends JiraWebActionSupport {
         String action = getHttpRequest().getParameter(PARAM_ACTION);
         if (action != null && action.equals("save")) {
             String newStoryPointField = getHttpRequest().getParameter(PARAM_STORY_POINT_FIELD);
+            scrumPokerSettingService.persistStoryPointField(newStoryPointField);
+
             String newSessionTimeout = getHttpRequest().getParameter(PARAM_SESSION_TIMEOUT);
-            scrumPokerSettingService.persistSettings(newStoryPointField, newSessionTimeout);
+            scrumPokerSettingService.persistSessionTimehout(Integer.valueOf(newSessionTimeout));
+
+            String newDefaultProjectActivation = getHttpRequest().getParameter(PARAM_DEFAULT_PROJECT_ACTIVATION);
+            scrumPokerSettingService.persistDefaultProjectActivation(Boolean.valueOf(newDefaultProjectActivation));
         }
         return "success";
     }