+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerSettingService.java
@@ -16,6 +16,13 @@ public interface ScrumPokerSettingService {
     String loadStoryPointField();
 
     /**
+     * Persist the story point field into the global settings.
+     *
+     * @param storyPointField story point field
+     */
+    void persistStoryPointField(String storyPointField);
+
+    /**
      * Loads the currently saved session timeout from the global settings.
      *
      * @return session timeout in hours
@@ -23,11 +30,24 @@ public interface ScrumPokerSettingService {
     Integer loadSessionTimeout();
 
     /**
-     * Persist the global settings of the plugin.
+     * Persist the session timeout into the global settings.
      *
-     * @param storyPointField story point field id
-     * @param sessionTimeout  session timeout in hours
+     * @param sessionTimeout session timehout in hours
      */
-    void persistSettings(String storyPointField, String sessionTimeout);
+    void persistSessionTimehout(Integer sessionTimeout);
+
+    /**
+     * Loads the currently saved default project activation from the global settings.
+     *
+     * @return default project activation
+     */
+    boolean loadDefaultProjectActivation();
+
+    /**
+     * Persist the defaultproject activation into the global settings.
+     *
+     * @param defaultProjectActivation default project activation
+     */
+    void persistDefaultProjectActivation(boolean defaultProjectActivation);
 
 }