+++ src/test/java/de/codescape/jira/plugins/scrumpoker/condition/ScrumPokerForIssueConditionTest.java
@@ -6,6 +6,7 @@ import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.plugin.webfragment.model.JiraHelper;
 import com.atlassian.jira.user.ApplicationUser;
 import de.codescape.jira.plugins.scrumpoker.service.EstimationFieldService;
+import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingService;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.InjectMocks;
@@ -32,6 +33,9 @@ public class ScrumPokerForIssueConditionTest {
     private EstimationFieldService estimationFieldService;
 
     @Mock
+    private ScrumPokerSettingService scrumPokerSettingService;
+
+    @Mock
     private ApplicationUser applicationUser;
 
     @Mock
@@ -50,10 +54,19 @@ public class ScrumPokerForIssueConditionTest {
     private Issue issue;
 
     @Test
-    public void shouldDisplayForEditableIssueWithStoryPointField() {
+    public void shouldDisplayForEditableIssueWithStoryPointFieldAndGloballyActiveScrumPoker() {
         expectThatIssueContainsTheStoryPointField();
         expectThatIssueIsEditable();
+        expectThatDefaultProjectActivationIsEnabled();
         assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, issue, jiraHelper), is(true));
+    }
+
+    @Test
+    public void shouldNotDisplayForEditableIssueWithStoryPointFieldAndGloballyInactiveScrumPoker() {
+        expectThatIssueContainsTheStoryPointField();
+        expectThatIssueIsEditable();
+        expectThatDefaultProjectActivationIsDisabled();
+        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, issue, jiraHelper), is(false));
     }
 
     @Test
@@ -94,6 +107,14 @@ public class ScrumPokerForIssueConditionTest {
 
     private void expectThatEstimationFieldExists() {
         when(estimationFieldService.findStoryPointField()).thenReturn(storyPointField);
+    }
+
+    private void expectThatDefaultProjectActivationIsEnabled() {
+        when(scrumPokerSettingService.loadDefaultProjectActivation()).thenReturn(true);
+    }
+
+    private void expectThatDefaultProjectActivationIsDisabled() {
+        when(scrumPokerSettingService.loadDefaultProjectActivation()).thenReturn(false);
     }
 
 }