+++ changelog.md
@@ -5,6 +5,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 ## [Unreleased]
 
 * docs: provide EULA and privacy statement in online documentation
+* feat: add global configuration to enable or disable Scrum Poker per default
 
 ## [3.7.0] - 2018-12-03